refactor(bookmark): update repository package to repositories; fix entity scan path in auto configuration, delete dto and mapper
