--- bookmark/src/main/java/pt/estga/bookmark/BookmarkDto.java
@@ -1,11 +1,0 @@
-package pt.estga.bookmark;
-
-import pt.estga.shared.enums.TargetType;
-
-public record BookmarkDto(
-        Long id,
-        TargetType type,
-        String targetId,
-        Object content
-) {}
-
