+++ bookmark/src/main/java/pt/estga/bookmark/entities/BaseBookmark.java
@@ -29,4 +29,3 @@ public abstract class BaseBookmark {
     @CreationTimestamp
     private Instant createdAt;
 }
-
